Move types needed in Dhall.DirectoryTree to own module
In order to fix the Windows builds observed in #2452 we only include the
`FromDhall CUid` and `FromDhall CGid` instance only if we are on that
platform. Those types are not available there which caused the CI
failure.

Fixes #2452
